6-size.c

+++ 0x00-hello_world/6-size.c
@@ -12,10 +12,10 @@ int main(void)
 	long long int d;
 	float f;
 
-pintf("Size of a char: %ld byte(s)\n", sizeof(a));
-pintf("Size of an int: %ld byte(s)\n", sizeof(b));
-pintf("Size of a long int: %ld byte(s)\n", sizeof(c));
-pintf("Size of a long long int: %ld byte(s)\n", sizeof(d));
-pintf("Size of a float: %ld byte(s)\n", sizeof(f));
-	return (0);
+pintf("Size of a char: %lu byte(s)\n", (unsigned long) sizeof(a));
+pintf("Size of an int: %lu byte(s)\n", (unsigned long) sizeof(b));
+pintf("Size of a long int: %lu byte(s)\n", (unsigned long) sizeof(c));
+pintf("Size of a long long int: %lu byte(s)\n", (unsigned long) sizeof(d));
+pintf("Size of a float: %lu byte(s)\n", (unsigned long) sizeof(f));
+return (0);
 }
